backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
alembic>=1.13,<2.0
pydantic-settings>=2.5,<3.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
numpy>=1.26,<3.0
pandas>=2.2,<3.0
pyarrow>=15,<22
pillow>=10,<12
scipy>=1.13,<2.0
implicit>=0.7,<1.0
scikit-learn>=1.5,<2.0
torch>=2.2,<3.0
transformers>=4.46,<5.0
faiss-cpu>=1.8,<2.0
